+++ toolkit/scaffold/project/notes/context.py
@@ -1,4 +1,0 @@
-NOTES_USER_INPUT_CONTEXT = {
-    "programing_language": "python",
-    "enable_library_module": True,
-}